# API-Azure/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de Gerenciamento de Recursos em Nuvem",
    "description": "API para gerenciar recursos em nuvem utilizando Flask",
    "version": "1.0"
  },
...
  "paths": {
    "/azure/login": {
      "post": {
        "summary": "Realiza o login na Azure",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Credenciais de login na Azure",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login na Azure realizado com sucesso"
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        },
        "tags": ["Azure"]
      }
    },
...
  }
}